Revert "Added button to show mediations"
This reverts commit 56776fdb032c59bb2d7dbd80ed2aca23bf3f1173.

This commit crashes.

Note that the commit before it, [43bad00 - Integrating Unity Ads],
crashed immediately on the device when I went back to it using git
if I didn't delete the Unity Library folder. (I've never had to do
that for going back in time before.) With a 'clean' reimport,
that revision works, but this one still fails.

I'm reverting this commit to push new work here that does work.